+++ cabal.project
@@ -22,7 +22,8 @@ source-repository-package
   type: git
   location: https://github.com/IntersectMBO/formal-ledger-specifications.git
   subdir: ledgerSrc/haskell/Ledger
-  tag: 79637b65e241537efbbad1a1125b7f7af8945b07
+  tag: 8e2dbbf5bc897ec89922a7e75e282d5bb5dda650
+  --sha256: sha256-jfphBcaojCyiHTsetEVqXzqkBxIo8UJ/GGOLD80iW+k=
   --sha256: sha256-L0oi+hFjFw8BE75jWtD+jAtBwaIDrMb8rqJGfBhSELE=
 -- NOTE: If you would like to update the above, look for the `haskell-package`
 -- branch in the `formal-ledger-specifications` repo where the HEAD contains
